django/requirements.txt
asgiref>=3.4.1
beautifulsoup4>=4.10.0
boto>=2.49.0
certifi>=2021.5.30
charset-normalizer>=2.0.6
colorama>=0.4.4
configparser>=5.0.2
crayons>=0.4.0
Django>=3.2.7
django-ses>=3.5.0
gunicorn>=21.0.0
future>=0.18.2
idna>=3.2
line-bot-sdk>=1.20.0
mysqlclient==2.1.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
pytz>=2021.1
requests>=2.26.0
selenium>=3.141.0
soupsieve>=2.2.1
sqlparse>=0.4.2
urllib3>=1.26.6
webdriver-manager>=3.4.2
